Modify sectionId field datatype from int to long

+++ backend/src/main/java/com/github/performancemonitor/model/dto/SectionPercentDto.java
@@ -15,7 +15,7 @@ public class SectionPercentDto {
 
     private int percent;
 
-    private int sectionId;
+    private long sectionId;
 
     private String sectionName;
